# api/docs/swagger.yaml
basePath: /api
definitions:
...
  books.AuthorSearchItem:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  books.BookDetailsResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/books.AuthorSearchItem'
        type: array
      id:
        type: string
      image_url:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.Error'
      meta: {}
    type: object
  response.PageMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  reviews.ReviewListingItem:
    properties:
      heart_score:
        type: number
      id:
        type: string
      poo_score:
        type: number
      text:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  reviews.ReviewListingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/reviews.ReviewListingItem'
        type: array
      meta:
        $ref: '#/definitions/response.PageMeta'
    type: object
  search.AuthorSearchItem:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  search.BookSearchItem:
    properties:
      authors:
        items:
          $ref: '#/definitions/search.AuthorSearchItem'
        type: array
      id:
        type: string
      image_url:
        type: string
      similarity:
        type: number
      title:
        type: string
    type: object
  search.BookSearchResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/search.BookSearchItem'
        type: array
      meta:
        $ref: '#/definitions/response.PageMeta'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API that will change the world
  title: BS Books API
  version: "1.0"
paths:
...
  /books/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve book details by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/books.BookDetailsResponse'
        "404":
          description: Book not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Book by ID
      tags:
      - Books
  /books/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Retrieve all reviews associated with a specific book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reviews.ReviewListingResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get reviews for a book
      tags:
      - reviews
  /books/search:
    get:
      consumes:
      - application/json
      description: Search for books by title
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/search.BookSearchResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Search Books
      tags:
      - Books
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"